added line note on teaching flexbox
Given that one of the learning outcomes for Week 1 is to build a site with a non-trivial layout, we *must* teach at least one form of CSS layout methodology.

Flexbox seems to be the best way to go, it's not as confusing as floats and virtually all browsers support the spec. Additionally, it's fairly easy to pick up, can be taught in about an hour or less.